Adding new features Changed versioning to match latest supported Laravel version Changed package name to match `Khooz`
Features:

* Added an sliding pivot for code generation functionality
* Added an sliding pivot and a range for code checking functionality
